+++ src/Areas/VirtualDesktop/Options/Hostpool/HostpoolListOptions.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.VirtualDesktop.Options.Hostpool
+{
+    public class HostpoolListOptions : SubscriptionOptions
+    {
+    }
+}